Moved FilterInput creation logic into a separate factory class.
