+++ src/main/java/com/checkmarx/sdk/service/CxService.java
@@ -116,11 +116,14 @@ public class CxService implements CxClient{
     public static final String FOUND_TEAM = "Found team {} with ID {}";
     public static final String ONLY_SUPPORTED_IN_90_PLUS = "Operation only supported in 9.0+";
     public static final String ERROR_GETTING_TEAMS = "Error occurred while retrieving Teams";
+    public static final String SCAN_CREATION_ERROR = "Error occurred while creating Scan for project {}, http error {}";
     private final CxProperties cxProperties;
     private final CxLegacyService cxLegacyService;
     private final CxAuthService authClient;
     private final RestTemplate restTemplate;
     private final ScanSettingsClient scanSettingsClient;
+
+    private final FilterInputFactory filterInputFactory;
     private final FilterValidator filterValidator;
 
     public CxService(CxAuthService authClient,
@@ -128,6 +131,7 @@ public class CxService implements CxClient{
                      CxLegacyService cxLegacyService,
                      @Qualifier("cxRestTemplate") RestTemplate restTemplate,
                      ScanSettingsClient scanSettingsClient,
+                     FilterInputFactory filterInputFactory,
                      FilterValidator filterValidator) {
         this.authClient = authClient;
         this.cxProperties = cxProperties;
@@ -135,17 +139,11 @@ public class CxService implements CxClient{
         this.restTemplate = restTemplate;
         this.scanSettingsClient = scanSettingsClient;
         this.filterValidator = filterValidator;
+        this.filterInputFactory = filterInputFactory;
     }
 
     /**
      * Create Scan for a projectId
-     *
-     * @param projectId
-     * @param incremental
-     * @param isPublic
-     * @param forceScan
-     * @param comment
-     * @return
      */
     public Integer createScan(Integer projectId, boolean incremental, boolean isPublic, boolean forceScan, String comment) {
         CxScan scan = CxScan.builder()
@@ -165,7 +163,7 @@ public class CxService implements CxClient{
             log.info("Scan created with Id {} for project Id {}", id, projectId);
             return Integer.parseInt(id);
         } catch (HttpStatusCodeException e) {
-            log.error("Error occurred while creating Scan for project {}, http error {}", projectId, e.getStatusCode());
+            log.error(SCAN_CREATION_ERROR, projectId, e.getStatusCode());
             log.error(ExceptionUtils.getStackTrace(e));
         }
         return UNKNOWN_INT;
@@ -191,7 +189,7 @@ public class CxService implements CxClient{
             log.info("Scan found with Id {} for project Id {}", id, projectId);
             return Integer.parseInt(id);
         } catch (HttpStatusCodeException e) {
-            log.error("Error occurred while creating Scan for project {}, http error {}", projectId, e.getStatusCode());
+            log.error(SCAN_CREATION_ERROR, projectId, e.getStatusCode());
             log.error(ExceptionUtils.getStackTrace(e));
         }
         return UNKNOWN_INT;
@@ -254,7 +252,7 @@ public class CxService implements CxClient{
                 }
             }
         } catch (HttpStatusCodeException e) {
-            log.error("Error occurred while creating Scan for project {}, http error {}", projectId, e.getStatusCode());
+            log.error(SCAN_CREATION_ERROR, projectId, e.getStatusCode());
             log.error(ExceptionUtils.getStackTrace(e));
         } catch (NullPointerException e) {
             log.error("Error parsing JSON response for dateAndTime status. {}", ExceptionUtils.getMessage(e));
@@ -265,9 +263,6 @@ public class CxService implements CxClient{
 
     /**
      * Get the status of a given scanId
-     *
-     * @param scanId
-     * @return
      */
     @Override
     public Integer getScanStatus(Integer scanId) {
@@ -291,7 +286,6 @@ public class CxService implements CxClient{
 
     /**
      * Generate a scan report request (xml) based on ScanId
-     * @return
      */
     @Override
     public Integer createScanReport(Integer scanId) {
@@ -318,7 +312,6 @@ public class CxService implements CxClient{
 
     /**
      * Get the status of a report being generated by reportId
-     * @return
      */
     @Override
     public Integer getReportStatus(Integer reportId) throws CheckmarxException{
@@ -470,8 +463,6 @@ public class CxService implements CxClient{
 
     /**
      * Retrieve the report by reportId, mapped to ScanResults DTO, applying filtering as requested
-     *
-     * @throws CheckmarxException
      */
     @Override
     public CxXMLResultsType getXmlReportContent(Integer reportId) throws CheckmarxException {
@@ -575,7 +566,7 @@ public class CxService implements CxClient{
     public Map<String, String> getCustomFields(Integer projectId) {
         Map<String, String> customFields = new HashMap<String, String>();
         log.info("Fetching custom fields from project ID ".concat(projectId.toString()));
-        CxProject cxProject = getProject(Integer.valueOf(projectId));
+        CxProject cxProject = getProject(projectId);
         if (cxProject != null) {
             for (CxProject.CustomField customField : cxProject.getCustomFields()) {
                 customFields.put(customField.getName(), customField.getValue());
@@ -646,9 +637,6 @@ public class CxService implements CxClient{
         }
     }
 
-    /**
-     * @throws CheckmarxException
-     */
     public ScanResults getOsaReportContent(File vulnsFile, File libsFile, List<Filter> filter) throws CheckmarxException {
         if (vulnsFile == null || libsFile == null) {
             throw new CheckmarxException("Files not provided for processing of OSA results");
@@ -749,20 +737,8 @@ public class CxService implements CxClient{
     }
 
 
-    private List<CxOsa> getOSAVulnsByLibId(List<CxOsa> osaVulns, String libId) {
-        List<CxOsa> vulns = new ArrayList<>();
-        for (CxOsa v : osaVulns) {
-            if (v.getLibraryId().equals(libId)) {
-                vulns.add(v);
-            }
-        }
-        return vulns;
-    }
-
-
     /**
      * @param filter determines which SAST findings will be mapped into XIssue-s.
-     * @param session
      * @param cxIssueList list that will be populated during this method execution.
      * @param cxResults SAST-specific scan results based on SAST XML report.
      */
@@ -776,7 +752,7 @@ public class CxService implements CxClient{
                 ScanResults.XIssue.XIssueBuilder xIssueBuilder = ScanResults.XIssue.builder();
                 /*Top node of each issue*/
                 for (ResultType resultType : result.getResult()) {
-                    FilterInput filterInput = FilterInput.getInstance(result, resultType);
+                    FilterInput filterInput = filterInputFactory.fromCxSastFinding(result, resultType);
                     if (filterValidator.passesFilter(filterInput, sastFilters)) {
                         boolean falsePositive = false;
                         if(!resultType.getFalsePositive().equalsIgnoreCase("FALSE")){
@@ -1019,8 +995,6 @@ public class CxService implements CxClient{
 
     /**
      * Get All Projects in Checkmarx
-     *
-     * @return
      */
     public List<CxProject> getProjects() throws CheckmarxException {
         HttpEntity httpEntity = new HttpEntity<>(authClient.createAuthHeaders());
@@ -1038,9 +1012,7 @@ public class CxService implements CxClient{
     }
 
     /**
-     * Get All Projects in Checkmarx
-     *
-     * @return
+     * Gets all CxSAST projects
      */
     public List<CxProject> getProjects(String teamId) throws CheckmarxException {
         HttpEntity httpEntity = new HttpEntity<>(authClient.createAuthHeaders());
@@ -1214,11 +1186,6 @@ public class CxService implements CxClient{
 
     /**
      * Set Repository details for a project
-     *
-     * @param projectId
-     * @param gitUrl
-     * @param branch
-     * @throws CheckmarxException
      */
     public void setProjectRepositoryDetails(Integer projectId, String gitUrl, String branch) throws CheckmarxException {
         CxProjectSource projectSource = CxProjectSource.builder()
@@ -1240,9 +1207,6 @@ public class CxService implements CxClient{
 
     /**
      * Update name and/or owning team for a project
-     *
-     * @param cxProject
-     * @throws CheckmarxException
      */
     public void updateProjectDetails(CxProject cxProject) throws CheckmarxException {
         String strJSON = "{'name':'%s','owningTeam':'%s'}";
@@ -1262,10 +1226,6 @@ public class CxService implements CxClient{
 
     /**
      * Upload file (zip of source) for a project
-     *
-     * @param projectId
-     * @param file
-     * @throws CheckmarxException
      */
     public void uploadProjectSource(Integer projectId, File file) throws CheckmarxException {
         HttpHeaders headers = authClient.createAuthHeaders();
@@ -1316,7 +1276,6 @@ public class CxService implements CxClient{
      *
      * @param teamPath Fully qualified name/path of the team to lookup
      * @return TeamID of the team, or UNKNOWN (-1)
-     * @throws CheckmarxException
      */
     public String getTeamId(String teamPath) throws CheckmarxException {
         HttpEntity httpEntity = new HttpEntity<>(authClient.createAuthHeaders());
@@ -1706,7 +1665,7 @@ public class CxService implements CxClient{
             log.info("Scan created with Id {} for project Id {}", id, projectId);
             return Integer.parseInt(id);
         } catch (HttpStatusCodeException e) {
-            log.error("Error occurred while creating Scan for project {}, http error {}", projectId, e.getStatusCode());
+            log.error(SCAN_CREATION_ERROR, projectId, e.getStatusCode());
             log.error(ExceptionUtils.getStackTrace(e));
         }
         log.info("...Finished creating scan");
